+++ app/src/main/java/org/sopt/stamp/data/remote/model/response/MissionResponse.kt
@@ -1,0 +1,12 @@
+package org.sopt.stamp.data.remote.model.response
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class MissionResponse(
+    val id: Int,
+    val title: String,
+    val level: Int,
+    val profileImage: List<String>,
+    val isCompleted: Boolean
+)
